Sarting framework for the project.
This uses an object oriented php library for the classes.  I included
bootstrap 4 with flexbox, bootstrap js, and jquery.  The way this is set
up is that you put pages in the pages directory (everything is php) and
the index.php will automatically fill in this page from the url.  Still
need to set up a database.  Please use bootstrap for pretty much
everything, docs are online.  Familiarize with the code.  Also please
tabs not spaces :100: